+++ staircase/__init__.py
@@ -10,3 +10,5 @@ from .stairs import _mean as mean
 from .stairs import _median as median
 from .stairs import _min as min
 from .stairs import _max as max
+
+from .test_data import make_test_data